same: Split clock_config.c in SAMD21 and SAMD51 files.
And put the into the variants tree. The file got a little bit long
and hard to read.
